Replace the Logger trait with a simple Vec<String> - API change
Previously all log messages were written to a Struct that implemented the Logger trait.
This allowed the caller of the a2lfile API to receive the messages while the operation was still in progress.
The idea was that this could be useful in a gui use case.
In reality no gui exists and the interface was over complicated.
It's not even necessary to get the messages asynchronously, because all operations are very fast.