make finding the end of the A2ML block more robust
Previously the first occurrence of "/end A2ML" was used as the ending of the A2ML block.
This could fail if there were extra spaces, e.g. "/end  A2ML".
Additionally this did not respect comments; if the string occured
in a comment, then the block was considered to end there.

Now the code skips over comments and only needs "/end" to mark the end of the block.