Replace the proc-macro in specification.rs with the generated code
The code was generated with cargo expand and then cleaned up by hand.
Unfortunately rust-analyzer is unable to handle the proc macro, and
emits false errors for code using a2lfile.

The original form of the file is preserved as specification_orig.rs,
and could be used or reinstated if needed.