Additional checks for RECORD_LAYOUT, AXIS_DESCR, limits
The RECORD_LAYOUT used by a CHARACTERISTIC must contain FNC_VALUES.
For each internal axis, it must contain an AXIS_PTS_X/Y/Z

There should be a specific number of AXIS_DESCR depending on the type
of the characteristic. For example, a VALUE has zero axes, while a MAP needs two.

The lower and upper limits of CHARACTERISTICs, MEASUREMENTs and AXIS_PTS
are checked for plausibility based on the data type of the item and its COMPU_METHOD