account for precision problems while comparing limits
The values in the a2l might not match the calculated limits exactly.
For example a file contained the limit 3.40282346639E38, while the
calculated limit was 3.4028234663852886E38.
The value in the file is slightly rounded up (on the final digit) compared
to the calculated value, but this should not trigger a warning.

To fix this, the comparison of limits now includes a 0.0001% tolerance,
which is big enough to hide rounding problems while still being small
enough to detect real problems.