maintain automatic name to object mapping for named objects
Named objects are now stored in ItemLists instead of Vecs.
An ItemList is overall like an IndexMap, but the key is the name inside the object.
The NameMap no longer needed, so it has been removed

The name property of various blocks that can be stored in itemlists is now protected, and
should be accessed though Object::get_name() or ItemList::rename_item()